#37049 Changing to tests easier to setup when validating number of optimizations

+++ Domain/Optimization/IntradayOptmizerLimiter.cs
@@ -16,7 +16,7 @@ namespace Teleopti.Ccc.Domain.Optimization
 		public bool CanJumpOutEarly(int totalNumberOfAgentsInSkillGroup, int optimizedNumberOfAgentsInSkillGroup)
 		{
 			return totalNumberOfAgentsInSkillGroup >= _minSizeLimit &&
-						 (optimizedNumberOfAgentsInSkillGroup/totalNumberOfAgentsInSkillGroup) >= _minPercentOfGroupLimit.Value;
+				((double)optimizedNumberOfAgentsInSkillGroup/totalNumberOfAgentsInSkillGroup) >= _minPercentOfGroupLimit.Value;
 		}
 	}
 }
